feat(alkaline): move to a Stream based Adapter API
Primarily this was done to allow state to exist across multiple
`Adapter::read()` calls, as might be required for tracking a
database Offset, external API page, or etc.
This state seemed essential for very basic impls.

An alternative design was experimented with where by the
Adapter would return a new associated type which could be mutated,
but this made the general requirements for Adapter more complex
for cases where it wasn't required.

Stream is reasonably easy for cases where this complexity is not
required, while being fairly powerful for cases that need mutated
state.